ai-verifier/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
scikit-learn==1.6.1
numpy==2.2.3
pandas==2.2.3
httpx==0.28.1
joblib==1.4.2
python-dotenv==1.0.1
